requirements.txt
paho-mqtt
Flask
Ask
flask-ask
